Always allow administrative editing of CAS username on user_profile_form.
Previously editing was only allowed if the user did not have multiple
associated CAS usernames. Now we allow editing of the first CAS username.
This functionality mimics editing URL aliases assigned to nodes.

User object also have an additional key of 'cas_name' which is the
first CAS username assigned to the account.