Update README to refer to correct Drupal 6 paths.
The README was previously copied from the Drupal 7 branch,
and so referred to incorrect URLs and page breadcrumbs.

Additionally, the UI in the mass user listing is slightly
different in D6 -- we cannot easily add a column to the
table so instead we put the CAS username(s) in parentheses.